스터디 상세 조회를 팩토리로 분리하라
스터디 상세조회 화면을 팩토리로 분리합니다
