+++ app/src/main/java/com/example/bookclub/controller/AccountController.java
@@ -2,7 +2,7 @@ package com.example.bookclub.controller;
 
 import com.example.bookclub.domain.account.Account;
 import com.example.bookclub.dto.StudyApiDto;
-import com.example.bookclub.repository.study.JpaStudyRepository;
+import com.example.bookclub.infrastructure.study.JpaStudyRepository;
 import com.example.bookclub.security.CurrentAccount;
 import com.example.bookclub.security.UserAccount;
 import org.springframework.security.access.AccessDeniedException;
