+++ app/src/main/java/com/example/bookclub/infrastructure/account/AccountRepositoryCustom.java
@@ -1,0 +1,13 @@
+package com.example.bookclub.infrastructure.account;
+
+import com.example.bookclub.domain.account.Account;
+
+import java.util.Optional;
+
+public interface AccountRepositoryCustom {
+    Optional<Account> findById(Long id);
+
+    Optional<Account> findByEmail(String email);
+
+	long getAllAccountsCount();
+}
