+++ app/src/main/java/com/example/bookclub/infrastructure/account/emailauthentication/JpaEmailAuthenticationRepository.java
@@ -1,0 +1,16 @@
+package com.example.bookclub.infrastructure.account.emailauthentication;
+
+import com.example.bookclub.domain.account.emailauthentication.EmailAuthentication;
+import com.example.bookclub.domain.account.emailauthentication.EmailAuthenticationRepository;
+import org.springframework.data.repository.CrudRepository;
+
+import java.util.Optional;
+
+public interface JpaEmailAuthenticationRepository
+        extends EmailAuthenticationRepository, CrudRepository<EmailAuthentication, Long> {
+    EmailAuthentication save(EmailAuthentication emailAuthentication);
+
+    Optional<EmailAuthentication> findByEmail(String email);
+
+    void delete(EmailAuthentication emailAuthentication);
+}
