+++ app/src/main/java/com/example/bookclub/infrastructure/study/studycomment/JpaStudyCommentRepository.java
@@ -1,0 +1,19 @@
+package com.example.bookclub.infrastructure.study.studycomment;
+
+import com.example.bookclub.domain.study.studycomment.StudyComment;
+import com.example.bookclub.domain.study.studycomment.StudyCommentRepository;
+import org.springframework.data.repository.CrudRepository;
+
+import java.util.List;
+import java.util.Optional;
+
+public interface JpaStudyCommentRepository
+		extends StudyCommentRepository, CrudRepository<StudyComment, Long> {
+	StudyComment save(StudyComment studyComment);
+
+	List<StudyComment> findByStudyId(Long id);
+
+	void delete(StudyComment studyComment);
+
+	Optional<StudyComment> findById(Long id);
+}
