--- app/src/main/java/com/example/bookclub/repository/account/AccountRepositoryCustom.java
@@ -1,13 +1,0 @@
-package com.example.bookclub.repository.account;
-
-import com.example.bookclub.domain.account.Account;
-
-import java.util.Optional;
-
-public interface AccountRepositoryCustom {
-    Optional<Account> findById(Long id);
-
-    Optional<Account> findByEmail(String email);
-
-	long getAllAccountsCount();
-}
