--- app/src/main/java/com/example/bookclub/repository/post/PostRepositoryCustom.java
@@ -1,9 +1,0 @@
-package com.example.bookclub.repository.post;
-
-import com.example.bookclub.domain.post.Post;
-
-import java.util.List;
-
-public interface PostRepositoryCustom {
-	List<Post> findByContent(String content);
-}
