+++ app/src/test/java/com/example/bookclub/application/AccountAuthenticationServiceTest.java
@@ -7,7 +7,7 @@ import com.example.bookclub.domain.account.role.RoleRepository;
 import com.example.bookclub.domain.study.Study;
 import com.example.bookclub.domain.uplodfile.UploadFile;
 import com.example.bookclub.common.exception.account.AccountEmailNotFoundException;
-import com.example.bookclub.repository.account.JpaAccountRepository;
+import com.example.bookclub.infrastructure.account.JpaAccountRepository;
 import com.example.bookclub.security.UserAccount;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
