+++ app/src/test/java/com/example/bookclub/application/AccountServiceTest.java
@@ -17,7 +17,7 @@ import com.example.bookclub.common.exception.account.AccountNicknameDuplicatedEx
 import com.example.bookclub.common.exception.account.AccountNotFoundException;
 import com.example.bookclub.common.exception.account.AccountPasswordBadRequestException;
 import com.example.bookclub.common.exception.account.emailauthentication.EmailNotAuthenticatedException;
-import com.example.bookclub.repository.account.JpaAccountRepository;
+import com.example.bookclub.infrastructure.account.JpaAccountRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
