+++ app/src/test/java/com/example/bookclub/application/LoginServiceTest.java
@@ -5,7 +5,7 @@ import com.example.bookclub.application.account.LoginService;
 import com.example.bookclub.domain.account.Account;
 import com.example.bookclub.dto.AccountDto;
 import com.example.bookclub.common.exception.account.AccountEmailNotFoundException;
-import com.example.bookclub.repository.account.JpaAccountRepository;
+import com.example.bookclub.infrastructure.account.JpaAccountRepository;
 import com.example.bookclub.security.UserAccount;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
