+++ app/src/test/java/com/example/bookclub/application/StudyServiceTest.java
@@ -22,8 +22,8 @@ import com.example.bookclub.common.exception.study.StudySizeFullException;
 import com.example.bookclub.common.exception.study.StudyStartAndEndDateNotValidException;
 import com.example.bookclub.common.exception.study.StudyStartAndEndTimeNotValidException;
 import com.example.bookclub.common.exception.study.StudyStartDateInThePastException;
-import com.example.bookclub.repository.account.JpaAccountRepository;
-import com.example.bookclub.repository.study.JpaStudyRepository;
+import com.example.bookclub.infrastructure.account.JpaAccountRepository;
+import com.example.bookclub.infrastructure.study.JpaStudyRepository;
 import com.example.bookclub.security.UserAccount;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
